+++ bootique-cayenne41-junit5/src/main/java/io/bootique/cayenne/v41/junit5/CayenneTester.java
@@ -24,6 +24,7 @@ import io.bootique.di.BQModule;
 import io.bootique.di.Binder;
 import io.bootique.junit5.BQTestScope;
 import io.bootique.junit5.scope.BQBeforeMethodCallback;
+import io.bootique.junit5.scope.BQBeforeScopeCallback;
 import org.apache.cayenne.Persistent;
 import org.apache.cayenne.configuration.server.ServerRuntime;
 import org.apache.cayenne.exp.Property;
@@ -42,7 +43,7 @@ import java.util.HashSet;
  *
  * @since 2.0
  */
-public class CayenneTester implements BQBeforeMethodCallback {
+public class CayenneTester implements BQBeforeScopeCallback, BQBeforeMethodCallback {
 
     private boolean refreshCayenneCaches;
     private boolean deleteBeforeEachTest;
@@ -272,10 +273,12 @@ public class CayenneTester implements BQBeforeMethodCallback {
     }
 
     @Override
-    public void beforeMethod(BQTestScope scope, ExtensionContext context) {
-
+    public void beforeScope(BQTestScope scope, ExtensionContext context) throws Exception {
         bootiqueHook.initIfNeeded();
+    }
 
+    @Override
+    public void beforeMethod(BQTestScope scope, ExtensionContext context) {
 
         if (refreshCayenneCaches) {
             getRuntimeManager().refreshCaches();