+++ bootique-cayenne-test/src/main/java/io/bootique/cayenne/test/CayenneModelUtils.java
@@ -40,6 +40,10 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Stream;
 
+/**
+ *  @deprecated since 3.0.M1, as we are we phasing out JUnit 4 support in favor of JUnit 5
+ */
+@Deprecated
 class CayenneModelUtils {
 
     static Table createTableModel(BQRuntime runtime, Class<?> entityType) {