+++ bootique-cayenne-test/src/main/java/io/bootique/cayenne/test/CayenneTestDataManagerBuilder.java
@@ -29,6 +29,10 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+/**
+ *  @deprecated since 3.0.M1, as we are we phasing out JUnit 4 support in favor of JUnit 5
+ */
+@Deprecated
 public class CayenneTestDataManagerBuilder {
 
     private BQRuntime runtime;