+++ bootique-cayenne41-test/src/main/java/io/bootique/cayenne/v41/test/CayenneModelUtils.java
@@ -36,6 +36,10 @@ import org.apache.cayenne.map.ObjRelationship;
 import java.util.*;
 import java.util.stream.Stream;
 
+/**
+ * @deprecated since 3.0.M1, as we are we phasing out JUnit 4 support in favor of JUnit 5
+ */
+@Deprecated
 class CayenneModelUtils {
 
     static Table createTableModel(BQRuntime runtime, Class<?> entityType) {