+++ bootique-cayenne41-test/src/main/java/io/bootique/cayenne/v41/test/SchemaListener.java
@@ -22,8 +22,11 @@ package io.bootique.cayenne.v41.test;
 import org.apache.cayenne.map.DataMap;
 
 /**
- * A listener that is called when the test DB schema is created.
+ * A listener that is called when the test DB schema is created
+ *
+ * @deprecated since 3.0.M1, as we are we phasing out JUnit 4 support in favor of JUnit 5
  */
+@Deprecated
 public interface SchemaListener {
 
     void afterSchemaCreated(DataMap dataMap);