+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #84 CayenneTester.assertQueryCount(int)
 * #85 Upgrade to Cayenne 4.2.M2
 * #86 NPE: ObjectContext access causes premature faulting of JUnit 5 DataSource
+* #88 Upgrade Cayenne 4.2 to M3
 
 ## 2.0.M1