+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+## 3.0.M1
+
+* #93 JUnit 4 support deprecation
+
 ## 2.0.B1
 
 * #84 CayenneTester.assertQueryCount(int)